requirements.txt
requests
beautifulsoup4
ebooklib
Pillow
tk